+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 
 - Changed process_mails output to filter easierly.
   [sgeulette]
+- Added new option `reset_flags` for `process_mails` to reset all flags of an email id.
+  [sgeulette]
 
 0.16 (2022-06-07)
 -----------------